rename CrawlerEngine, create CrawlerEngine interface, create new SimpleCrawlerEngine class
