+++ packages/crawler/src/crawler.ts
@@ -3,13 +3,13 @@
 import { Container } from 'inversify';
 import { registerCrawlerRunOptions } from './setup-crawler-container';
 import { CrawlerRunOptions } from './types/crawler-run-options';
-import { CrawlerEngine } from './crawler/crawler-engine';
+import { PuppeteerCrawlerEngine } from './crawler/puppeteer-crawler-engine';
 
 export class Crawler {
     constructor(private readonly container: Container) {}
 
     public async crawl(crawlerRunOptions: CrawlerRunOptions): Promise<void> {
         registerCrawlerRunOptions(this.container, crawlerRunOptions);
-        await this.container.get(CrawlerEngine).start(crawlerRunOptions);
+        await this.container.get(PuppeteerCrawlerEngine).start(crawlerRunOptions);
     }
 }
