+++ packages/crawler/src/index.ts
@@ -9,3 +9,5 @@ export { Crawler } from './crawler';
 export { setupCrawlerContainer } from './setup-crawler-container';
 export * from './level-storage/storage-documents';
 export { DbScanResultReader } from './scan-result-providers/db-scan-result-reader';
+export { CrawlerEngine } from './crawler/crawler-engine';
+export { SimpleCrawlerEngine } from './crawler/simple-crawler-engine';
